Refactor and reorganize tests
